collapse_pivot_names {solrium}	R Documentation

Collapse Pivot Field and Value Columns

Description

Convert a table consisting of columns in sets of 3 into 2 columns assuming that the first column of every set of 3 (field) is duplicated throughout all rows and should be removed. This type of structure is usually returned by facet.pivot responses.

Usage

collapse_pivot_names(data)

Arguments

data
a data.frame with every 2 columns representing a field and value and the final representing a count

Value

a data.frame

[Package solrium version 1.0.2 Index]