collection_clusterprop {solrium}	R Documentation

Add, edit, delete a cluster-wide property

Description

Important: whether add, edit, or delete is used is determined by the value passed to the val parameter. If the property name is new, it will be added. If the property name exists, and the value is different, it will be edited. If the property name exists, and the value is NULL or empty the property is deleted (unset).

Usage

collection_clusterprop(conn, name, val, raw = FALSE, callopts = list())

Arguments

conn
A solrium connection object, see SolrClient

name
(character) Name of the core or collection

val
(character) Required. The value of the property. If the value is empty or null, the property is unset.

raw
(logical) If TRUE, returns raw data in format specified by wt param

callopts
curl options passed on to crul::HttpClient

Examples

## Not run: 
(conn <- SolrClient$new())

# add the value https to the property urlScheme
collection_clusterprop(conn, name = "urlScheme", val = "https")

# status again
collection_clusterstatus(conn)$cluster$properties

# delete the property urlScheme by setting val to NULL or a 0 length string
collection_clusterprop(conn, name = "urlScheme", val = "")

## End(Not run)

[Package solrium version 1.0.2 Index]