collection_clusterstatus {solrium}	R Documentation

Get cluster status

Description

Fetch the cluster status including collections, shards, replicas, configuration name as well as collection aliases and cluster properties.

Usage

collection_clusterstatus(conn, name = NULL, shard = NULL,
  raw = FALSE, ...)

Arguments

conn
A solrium connection object, see SolrClient

name
(character) The name of the core to be created. Required

shard
(character) The shard(s) for which information is requested. Multiple shard names can be specified as a character vector.

raw
(logical) If TRUE, returns raw data

...
You can pass in parameters like property.name=value to set core property name to value. See the section Defining core.properties for details on supported properties and values. (https://lucene.apache.org/solr/guide/7_0/defining-core-properties.html)

Examples

## Not run: 
(conn <- SolrClient$new())
conn$collection_clusterstatus()
res <- conn$collection_clusterstatus()
res$responseHeader
res$cluster
res$cluster$collections
res$cluster$collections$gettingstarted
res$cluster$live_nodes

## End(Not run)

[Package solrium version 1.0.2 Index]