collection_createalias {solrium}	R Documentation

Create an alias for a collection

Description

Create a new alias pointing to one or more collections. If an alias by the same name already exists, this action will replace the existing alias, effectively acting like an atomic "MOVE" command.

Usage

collection_createalias(conn, alias, collections, raw = FALSE,
  callopts = list())

Arguments

conn
A solrium connection object, see SolrClient

alias
(character) Required. The alias name to be created

collections
(character) Required. A character vector of collections to be aliased

raw
(logical) If TRUE, returns raw data

callopts
curl options passed on to HttpClient

Examples

## Not run: 
(conn <- SolrClient$new())

if (!conn$collection_exists("thingsstuff")) {
  conn$collection_create(name = "thingsstuff")
}

conn$collection_createalias("tstuff", "thingsstuff")
conn$collection_clusterstatus()$cluster$collections$thingsstuff$aliases

## End(Not run)

[Package solrium version 1.0.2 Index]