collection_deletereplica {solrium}	R Documentation

Delete a replica

Description

Delete a replica from a given collection and shard. If the corresponding core is up and running the core is unloaded and the entry is removed from the clusterstate. If the node/core is down , the entry is taken off the clusterstate and if the core comes up later it is automatically unregistered.

Usage

collection_deletereplica(conn, name, shard = NULL, replica = NULL,
  onlyIfDown = FALSE, raw = FALSE, callopts = list(), ...)

Arguments

conn
A solrium connection object, see SolrClient

name
(character) Required. The name of the collection.

shard
(character) Required. The name of the shard that includes the replica to be removed.

replica
(character) Required. The name of the replica to remove.

onlyIfDown
(logical) When TRUE will not take any action if the replica is active. Default: FALSE

raw
(logical) If TRUE, returns raw data

callopts
curl options passed on to crul::HttpClient

...
You can pass in parameters like property.name=value to set core property name to value. See the section Defining core.properties for details on supported properties and values. (https://cwiki.apache.org/confluence/display/solr/Defining+core.properties)

Examples

## Not run: 
(conn <- SolrClient$new())

# create collection
if (!conn$collection_exists("foobar2")) {
  conn$collection_create(name = "foobar2", maxShardsPerNode = 2)
}

# status
conn$collection_clusterstatus()$cluster$collections$foobar2$shards$shard1

# add replica
conn$collection_addreplica(name = "foobar2", shard = "shard1")

# delete replica
## get replica name
nms <- names(conn$collection_clusterstatus()$cluster$collections$foobar2$shards$shard1$replicas)
conn$collection_deletereplica(name = "foobar2", shard = "shard1", replica = nms[1])

# status again
conn$collection_clusterstatus()$cluster$collections$foobar2$shards$shard1

## End(Not run)

[Package solrium version 1.0.2 Index]