collection_removerole {solrium}	R Documentation

Remove a role from a node

Description

Remove an assigned role. This API is used to undo the roles assigned using collection_addrole

Usage

collection_removerole(conn, role = "overseer", node, raw = FALSE, ...)

Arguments

conn
A solrium connection object, see SolrClient

role
(character) Required. The name of the role. The only supported role as of now is overseer (set as default).

node
(character) Required. The name of the node.

raw
(logical) If TRUE, returns raw data

...
You can pass in parameters like property.name=value to set core property name to value. See the section Defining core.properties for details on supported properties and values. (https://lucene.apache.org/solr/guide/7_0/defining-core-properties.html)

Examples

## Not run: 
(conn <- SolrClient$new())

# get list of nodes
nodes <- conn$collection_clusterstatus()$cluster$live_nodes
conn$collection_addrole(node = nodes[1])
conn$collection_removerole(node = nodes[1])

## End(Not run)

[Package solrium version 1.0.2 Index]