collection_requeststatus {solrium}	R Documentation

Get request status

Description

Request the status of an already submitted Asynchronous Collection API call. This call is also used to clear up the stored statuses.

Usage

collection_requeststatus(conn, requestid, raw = FALSE, ...)

Arguments

conn
A solrium connection object, see SolrClient

requestid
(character) Required. The user defined request-id for the request. This can be used to track the status of the submitted asynchronous task. -1 is a special request id which is used to cleanup the stored states for all of the already completed/failed tasks.

raw
(logical) If TRUE, returns raw data

...
You can pass in parameters like property.name=value to set core property name to value. See the section Defining core.properties for details on supported properties and values. (https://lucene.apache.org/solr/guide/7_0/defining-core-properties.html)

[Package solrium version 1.0.2 Index]