commit {solrium}	R Documentation

Commit

Description

Commit

Usage

commit(conn, name, expunge_deletes = FALSE, wait_searcher = TRUE,
  soft_commit = FALSE, wt = "json", raw = FALSE, ...)

Arguments

conn
A solrium connection object, see SolrClient

name
(character) A collection or core name. Required.

expunge_deletes
merge segments with deletes away. Default: FALSE

wait_searcher
block until a new searcher is opened and registered as the main query searcher, making the changes visible. Default: TRUE

soft_commit
perform a soft commit - this will refresh the 'view' of the index in a more performant manner, but without "on-disk" guarantees. Default: FALSE

wt
(character) One of json (default) or xml. If json, uses jsonlite::fromJSON() to parse. If xml, uses xml2::read_xml() to parse

raw
(logical) If TRUE, returns raw data in format specified by wt param

...
curl options passed on to crul::HttpClient

References

<>

Examples

## Not run: 
(conn <- SolrClient$new())

conn$commit("gettingstarted")
conn$commit("gettingstarted", wait_searcher = FALSE)

# get xml back
conn$commit("gettingstarted", wt = "xml")
## raw xml
conn$commit("gettingstarted", wt = "xml", raw = TRUE)

## End(Not run)

[Package solrium version 1.0.2 Index]