config_get {solrium}	R Documentation

Get Solr configuration details

Description

Get Solr configuration details

Usage

config_get(conn, name, what = NULL, wt = "json", raw = FALSE, ...)

Arguments

conn
A solrium connection object, see SolrClient

name
(character) The name of the core. If not given, all cores.

what
(character) What you want to look at. One of solrconfig or schema. Default: solrconfig

wt
(character) One of json (default) or xml. Data type returned. If json, uses fromJSON to parse. If xml, uses read_xml to parse.

raw
(logical) If TRUE, returns raw data in format specified by wt

...
curl options passed on to crul::HttpClient

Details

Note that if raw=TRUE, what is ignored. That is, you get all the data when raw=TRUE.

Value

A list, xml_document, or character

Examples

## Not run: 
# start Solr with Cloud mode via the schemaless eg: bin/solr -e cloud
# you can create a new core like: bin/solr create -c corename
# where <corename> is the name for your core - or creaate as below

# connect
(conn <- SolrClient$new())

# all config settings
conn$config_get("gettingstarted")

# just znodeVersion
conn$config_get("gettingstarted", "znodeVersion")

# just znodeVersion
conn$config_get("gettingstarted", "luceneMatchVersion")

# just updateHandler
conn$config_get("gettingstarted", "updateHandler")

# just updateHandler
conn$config_get("gettingstarted", "requestHandler")

## Get XML
conn$config_get("gettingstarted", wt = "xml")
conn$config_get("gettingstarted", "updateHandler", wt = "xml")
conn$config_get("gettingstarted", "requestHandler", wt = "xml")

## Raw data - what param ignored when raw=TRUE
conn$config_get("gettingstarted", raw = TRUE)

## End(Not run)

[Package solrium version 1.0.2 Index]