core_rename {solrium}	R Documentation

Rename a core

Description

Rename a core

Usage

core_rename(conn, name, other, async = NULL, raw = FALSE,
  callopts = list())

Arguments

conn
A solrium connection object, see SolrClient

name
(character) The name of the core to be created. Required

other
(character) The new name of the core. Required.

async
(character) Request ID to track this action which will be processed asynchronously

raw
(logical) If TRUE, returns raw data

callopts
curl options passed on to crul::HttpClient

Examples

## Not run: 
# start Solr with Schemaless mode via the schemaless eg:
#   bin/solr start -e schemaless
# you can create a new core like: bin/solr create -c corename
# where <corename> is the name for your core - or creaate as below

# connect
(conn <- SolrClient$new())

# Status of particular cores
path <- "~/solr-7.0.0/server/solr/testcore/conf"
dir.create(path, recursive = TRUE)
files <- list.files(
"~/solr-7.0.0/server/solr/configsets/sample_techproducts_configs/conf/",
full.names = TRUE)
invisible(file.copy(files, path, recursive = TRUE))
conn$core_create("testcore") # or create in CLI: bin/solr create -c testcore

# rename
conn$core_rename("testcore", "newtestcore")
## status
conn$core_status("testcore") # core missing
conn$core_status("newtestcore", FALSE) # not missing

# cleanup
conn$core_unload("newtestcore")

## End(Not run)

[Package solrium version 1.0.2 Index]