core_requeststatus {solrium}	R Documentation

Request status of asynchronous CoreAdmin API call

Description

Request status of asynchronous CoreAdmin API call

Usage

core_requeststatus(conn, requestid, raw = FALSE, callopts = list())

Arguments

conn
A solrium connection object, see SolrClient

requestid
The name of one of the cores to be removed. Required

raw
(logical) If TRUE, returns raw data

callopts
curl options passed on to crul::HttpClient

Examples

## Not run: 
# start Solr with Schemaless mode via the schemaless eg:
#   bin/solr start -e schemaless

# FIXME: not tested yet...
# (conn <- SolrClient$new())
# conn$core_requeststatus(requestid = 1)

## End(Not run)

[Package solrium version 1.0.2 Index]