solr_optimize {solrium}	R Documentation

Optimize

Description

Optimize

Usage

solr_optimize(conn, name, max_segments = 1, wait_searcher = TRUE,
  soft_commit = FALSE, wt = "json", raw = FALSE, ...)

Arguments

conn
A solrium connection object, see SolrClient

name
(character) A collection or core name. Required.

max_segments
optimizes down to at most this number of segments. Default: 1

wait_searcher
block until a new searcher is opened and registered as the main query searcher, making the changes visible. Default: TRUE

soft_commit
perform a soft commit - this will refresh the 'view' of the index in a more performant manner, but without "on-disk" guarantees. Default: FALSE

wt
(character) One of json (default) or xml. If json, uses jsonlite::fromJSON() to parse. If xml, uses xml2::read_xml() to parse

raw
(logical) If TRUE, returns raw data in format specified by wt param

...
curl options passed on to crul::HttpClient

Examples

## Not run: 
(conn <- SolrClient$new())

solr_optimize(conn, "gettingstarted")
solr_optimize(conn, "gettingstarted", max_segments = 2)
solr_optimize(conn, "gettingstarted", wait_searcher = FALSE)

# get xml back
solr_optimize(conn, "gettingstarted", wt = "xml")
## raw xml
solr_optimize(conn, "gettingstarted", wt = "xml", raw = TRUE)

## End(Not run)

[Package solrium version 1.0.2 Index]